refactor(oas3): fold validating required keys into parseObject
